feat: ported over parsing report descriptor
